pyproject.toml
[tool.poetry]
name = "pytest-markdown-docs"
version = "0.2.0"
description = "Run markdown code fences through pytest"
readme = "README.md"
authors = ["Modal Labs", "Elias Freider <elias@modal.com>"]
license = "MIT"
classifiers = ["Framework :: Pytest"]
repository = "https://github.com/modal-com/pytest-markdown-docs"
include = ["LICENSE"]

[tool.poetry.dependencies]
python = "^3.6"
markdown-it-py = "~=1.1.0"

[tool.poetry.dev-dependencies]
pytest = "~=6.2.5"

[tool.poetry.plugins]

[tool.poetry.plugins."pytest11"]
"pytest_markdown_docs" = "pytest_markdown_docs.plugin"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
